[fix]: 사용자 트랜드 수집 오류 수정

+++ domain/src/main/kotlin/ac/dnd/mour/android/domain/model/feature/statistics/GroupStatistics.kt
@@ -2,12 +2,12 @@ package ac.dnd.mour.android.domain.model.feature.statistics
 
 data class GroupStatistics(
     val event: String,
-    val amount: Long
+    val amount: Double
 ) {
     companion object {
         val empty = GroupStatistics(
             event = "",
-            amount = 0
+            amount = 0.0
         )
     }
 }
